Refactor storage multipart
